Fix unsigned current_frame_seen variable in frame decoders

+++ data/frameReceiver/src/PercivalEmulatorFrameDecoder.cpp
@@ -91,7 +91,7 @@ void PercivalEmulatorFrameDecoder::process_packet_header(size_t bytes_received, 
         LOG4CXX_INFO(packet_logger_, ss.str());
     }
 
-	uint32_t frame = get_frame_number();
+	int frame = static_cast<int>(get_frame_number());
 	uint16_t packet_number = get_packet_number();
 	uint8_t  subframe = get_subframe_number();
 	uint8_t  type = get_packet_type();
@@ -240,10 +240,10 @@ void PercivalEmulatorFrameDecoder::monitor_buffers(void)
     gettime(&current_time);
 
     // Loop over frame buffers currently in map and check their state
-    std::map<uint32_t, int>::iterator buffer_map_iter = frame_buffer_map_.begin();
+    std::map<int, int>::iterator buffer_map_iter = frame_buffer_map_.begin();
     while (buffer_map_iter != frame_buffer_map_.end())
     {
-        uint32_t frame_num = buffer_map_iter->first;
+        int frame_num = buffer_map_iter->first;
         int      buffer_id = buffer_map_iter->second;
         void*    buffer_addr = buffer_manager_->get_buffer_address(buffer_id);
         PercivalEmulator::FrameHeader* frame_header = reinterpret_cast<PercivalEmulator::FrameHeader*>(buffer_addr);
